feat: RoomData를 분리
